Refactor InputLoopService into Program.cs for CosmosDB simple sample
